+++ example/activatory_example.dart
@@ -4,18 +4,18 @@ main() {
   var activatory = new Activatory();
 
   // We can create primitive types
-  print('Int: ${activatory.getTyped<int>()}');
-  print('String: ${activatory.getTyped<String>()}');
-  print('DateTime: ${activatory.getTyped<DateTime>()}');
-  print('Bool: ${activatory.getTyped<bool>()}');
-  print('TestEnum: ${activatory.getTyped<TestEnum>()}');
+  print('Int: ${activatory.get<int>()}');
+  print('String: ${activatory.get<String>()}');
+  print('DateTime: ${activatory.get<DateTime>()}');
+  print('Bool: ${activatory.get<bool>()}');
+  print('TestEnum: ${activatory.get<TestEnum>()}');
 
   //We can create complex types
-  var complexClassInstance = activatory.getTyped<SomeComplexClass>();
+  var complexClassInstance = activatory.get<SomeComplexClass>();
   print('Complex.String: ${complexClassInstance.stringField}');
 
   //And yes, we can create them recursively
-  var moreComplexClassInstance = activatory.getTyped<MoreComplexClass>();
+  var moreComplexClassInstance = activatory.get<MoreComplexClass>();
   print('MoreComplex.DateTime: ${moreComplexClassInstance.dateTimeField}');
   print('MoreComplex.Complex.String: ${moreComplexClassInstance.someComplexClass.stringField}');
 
@@ -27,6 +27,7 @@ class MoreComplexClass {
   SomeComplexClass _someComplexField;
 
   MoreComplexClass(this._dateTimeField, this._someComplexField);
+
   DateTime get dateTimeField => _dateTimeField;
 
   SomeComplexClass get someComplexClass => _someComplexField;
@@ -34,6 +35,7 @@ class MoreComplexClass {
 
 class SomeComplexClass {
   String _stringField;
+
   SomeComplexClass(this._stringField);
 
   String get stringField => _stringField;