+++ lib/src/backends/params_object_backend.dart
@@ -1,6 +1,6 @@
 import 'package:activatory/src/activation_context.dart';
 import 'package:activatory/src/backends/generator_backend.dart';
-import 'package:activatory/src/params_object.dart';
+import 'package:activatory/src/params_object/params_object.dart';
 
 class ParamsObjectBackend<TValue> implements GeneratorBackend<TValue> {
   TValue get(ActivationContext context) {