+++ lib/src/params_object/null_value.dart
@@ -1,0 +1,5 @@
+import 'package:activatory/src/params_object/value.dart';
+
+class NullValue<T> extends Value<T> {
+  const NullValue() : super(null);
+}