+++ lib/src/params_object/params_object.dart
@@ -1,10 +1,5 @@
 import 'package:activatory/src/activation_context.dart';
-
-Value<T> v<T>(T value) => new Value<T>(value);
-
-class NullValue<T> extends Value<T> {
-  const NullValue() : super(null);
-}
+import 'package:activatory/src/params_object/value.dart';
 
 abstract class Params<TResult> {
   T get<T>(Value<T> value, ActivationContext ctx) {
@@ -15,10 +10,4 @@ abstract class Params<TResult> {
   }
 
   TResult resolve(ActivationContext ctx);
-}
-
-class Value<T> {
-  final T value;
-
-  const Value(this.value);
 }