+++ lib/src/params_object/value.dart
@@ -1,0 +1,5 @@
+class Value<T> {
+  final T value;
+
+  const Value(this.value);
+}