+++ test/test_classes.dart
@@ -1,4 +1,5 @@
 import 'package:activatory/activatory.dart';
+import 'package:activatory/params_object.dart';
 
 abstract class AbstractClass {
   int _intField;
@@ -17,11 +18,11 @@ class CtorWithTwoStringArgs {
   CtorWithTwoStringArgs(this._a, this._b);
 
   String get a => _a;
+
   String get b => _b;
 }
 
-class DefaultCtor {
-}
+class DefaultCtor {}
 
 class DefaultNamedValues {
   static const String defaultValue = "defaultValue is not generated";
@@ -32,11 +33,12 @@ class DefaultNamedValues {
   PrimitiveComplexObject _nullSetObject;
 
   String _notNullSetString;
-  DefaultNamedValues({
-    PrimitiveComplexObject notSetObject,
-    PrimitiveComplexObject nullSetObject = null,
-    String nullSetString = null,
-    String notNullSetString=defaultValue}) {
+
+  DefaultNamedValues(
+      {PrimitiveComplexObject notSetObject,
+      PrimitiveComplexObject nullSetObject = null,
+      String nullSetString = null,
+      String notNullSetString = defaultValue}) {
     _nullSetString = nullSetString;
     _notSetObject = notSetObject;
     _nullSetObject = nullSetObject;
@@ -44,9 +46,11 @@ class DefaultNamedValues {
   }
 
   PrimitiveComplexObject get notSetObject => _notSetObject;
+
   PrimitiveComplexObject get nullSetObject => _nullSetObject;
 
   String get nullSetString => _nullSetString;
+
   String get notNullSetString => _notNullSetString;
 }
 
@@ -57,12 +61,18 @@ class DefaultPositionalValues {
   final PrimitiveComplexObject _notSetObject;
   final PrimitiveComplexObject _nullSetObject;
 
-  DefaultPositionalValues([this._notSetObject, this._nullSetObject = null, this._nullSetString = null, this._notNullString = defaultStringValue]);
+  DefaultPositionalValues(
+      [this._notSetObject,
+      this._nullSetObject = null,
+      this._nullSetString = null,
+      this._notNullString = defaultStringValue]);
 
   PrimitiveComplexObject get notSetObject => _notSetObject;
+
   PrimitiveComplexObject get nullSetObject => _nullSetObject;
 
   String get nullSetStringValue => _nullSetString;
+
   String get notNullSetStringValue => _notNullString;
 }
 
@@ -73,6 +83,7 @@ class FactoryCtor {
   factory FactoryCtor(String stringField) {
     return new FactoryCtor._internal(stringField, null);
   }
+
   FactoryCtor._internal(this._stringField, this._nonFactoryField);
 
   String get nonFactoryField => _nonFactoryField;
@@ -87,6 +98,7 @@ class FactoryWithFixedValues {
   static final FactoryWithFixedValues c = new FactoryWithFixedValues._("C");
 
   String _field;
+
   factory FactoryWithFixedValues(String type) {
     switch (type) {
       case 'A':
@@ -99,6 +111,7 @@ class FactoryWithFixedValues {
         throw new ArgumentError(type);
     }
   }
+
   FactoryWithFixedValues._(this._field);
 
   String get field => _field;
@@ -150,6 +163,7 @@ class LinkedNode {
 
 class NamedCtor {
   String _stringField;
+
   NamedCtor.nonDefaultName(this._stringField);
 
   String get stringField => _stringField;
@@ -157,6 +171,7 @@ class NamedCtor {
 
 class NamedCtorsAndConstCtor {
   final String _field;
+
   const NamedCtorsAndConstCtor(this._field);
 
   NamedCtorsAndConstCtor.A() : this('A');
@@ -164,11 +179,13 @@ class NamedCtorsAndConstCtor {
   NamedCtorsAndConstCtor.B() : this('B');
 
   NamedCtorsAndConstCtor.C() : this('C');
+
   String get field => _field;
 }
 
 class NamedCtorsAndDefaultCtor {
   final String _field;
+
   NamedCtorsAndDefaultCtor(this._field);
 
   NamedCtorsAndDefaultCtor.createA() : this("A");
@@ -184,6 +201,7 @@ class NamedCtorsAndDefaultCtor {
 
 class NamedCtorsAndFactory {
   String _field;
+
   factory NamedCtorsAndFactory(String arg) {
     return NamedCtorsAndFactory._internal(arg);
   }
@@ -204,6 +222,7 @@ class NonPrimitiveComplexObject {
   int _intField;
 
   NonPrimitiveComplexObject(this._primitiveComplexObject, this._intField);
+
   int get intField => _intField;
 
   PrimitiveComplexObject get primitiveComplexObject => _primitiveComplexObject;
@@ -211,6 +230,7 @@ class NonPrimitiveComplexObject {
 
 class NoPublicCtor {
   String _stringField;
+
   NoPublicCtor._internal(this._stringField);
 
   String get stringField => _stringField;
@@ -228,18 +248,27 @@ class PrimitiveComplexObject {
       this._intField, this._stringField, this._doubleField, this._boolField, this._dateTimeField, this._enumField);
 
   bool get boolField => _boolField;
+
   DateTime get dateTimeField => _dateTimeField;
+
   double get doubleField => _doubleField;
+
   TestEnum get enumField => _enumField;
+
   int get intField => _intField;
+
   String get stringField => _stringField;
 }
 
 abstract class Task {
   DateTime get dueDate;
+
   int get id;
+
   bool get isRecurrent;
+
   bool get isTemplate;
+
   String get title;
 }
 
@@ -253,7 +282,7 @@ class TreeNode {
   List<TreeNode> get children => _children;
 }
 
-class PrimitiveIterableInCtor{
+class PrimitiveIterableInCtor {
   final Iterable<String> _field;
 
   PrimitiveIterableInCtor(this._field);
@@ -261,7 +290,7 @@ class PrimitiveIterableInCtor{
   Iterable<String> get field => _field;
 }
 
-class ComplexIterableInCtor{
+class ComplexIterableInCtor {
   final Iterable<PrimitiveIterableInCtor> _field;
 
   ComplexIterableInCtor(this._field);
@@ -269,16 +298,13 @@ class ComplexIterableInCtor{
   Iterable<PrimitiveIterableInCtor> get field => _field;
 }
 
-abstract class ParentClass{
+abstract class ParentClass {}
 
-}
-
-class ChildClass extends ParentClass{
-
-}
+class ChildClass extends ParentClass {}
 
 class FiledsWithPublicSetters {
   final String _finalField;
+
   String get finalField => _finalField;
 
   FiledsWithPublicSetters(this._finalField);